[refactor][move] creates a library/ dir and appropriately organizes there (got bigger than I thought, oops, should've started this way)
