+++ library/Announcer.js
